finish

+++ src/services/contacts.js
@@ -9,11 +9,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
   }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactsCollection.find();
+  const contactsQuery = ContactsCollection.find(userId,);
   if (filter.type) {
     contactsQuery.where('contactType').equals(filter.type);
   }
@@ -22,7 +23,7 @@ export const getAllContacts = async ({
     contactsQuery.where('isFavourite').equals(filter.isFavourite);
   }
 
-   const [contactsCount, contacts] = await Promise.all([
+    const [contactsCount, contacts] = await Promise.all([
     ContactsCollection.find().merge(contactsQuery).countDocuments(),
     contactsQuery
       .skip(skip)
@@ -40,8 +41,8 @@ export const getAllContacts = async ({
 };
 
 
-export const getContactById = async (contactId) => {
-  const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async ({ contactId, userId }) => {
+  const contact = await ContactsCollection.findById({_id: contactId, userId: userId,  });
   return contact;
 };
 
@@ -50,14 +51,13 @@ export const createContact = async (payload) => {
   return contact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
-  const rawResult = await ContactsCollection.findOneAndUpdate(
-    { _id: contactId },
+export const updateContact = async ({ contactId, userId}, payload, options = {}) => {
+    const rawResult = await ContactsCollection.findOneAndUpdate(
+    { _id: contactId, userId: userId },
     payload,
     {
-      new: true,
-      includeResultMetadata: true,
       ...options,
+      includeResultMetadata: true,
     },
   );
 
@@ -69,7 +69,7 @@ export const updateContact = async (contactId, payload, options = {}) => {
   };
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async ({ contactId, userId }) => {
+  const contact = await ContactsCollection.findOneAndDelete({ _id: contactId , userId: userId});
   return contact;
 };
